+++ src/components/index.js
@@ -1,8 +1,7 @@
 export { default as ActionItem } from './ActionItem';
+export { default as BoundPiecesList } from './BoundPiecesList';
 export { default as FilterNavigation } from './FilterNavigation';
-export { default as InstanceFilters, InstanceFiltersBrowse, instanceFilterRenderer } from './InstanceFilters';
-export { default as HoldingsRecordFilters, holdingsRecordFilterRenderer } from './HoldingsRecordFilters';
-export { default as ItemFilters, itemFilterRenderer } from './ItemFilters';
+export { default as InstanceFiltersBrowse } from './InstanceFiltersBrowse';
 export { default as ItemViewSubheader } from './ItemViewSubheader';
 export { default as InstancePlugin } from './InstancePlugin';
 export { default as ConnectedTitle } from './ConnectedTitle';